Update file structure and move to o-forms v3 mixins
