+++ main.scss
@@ -5,5 +5,8 @@
 @import "o-forms/main";
 @import "o-colors/main";
 
-@import "./src/email-only-signup";
+@import "./src/scss/color-use-cases";
+@import "./src/scss/mixins";
+
+@import "./src/scss/email-only-signup";
