+++ src/js/email-only-signup.js
