+++ src/js/helpers.js
