+++ src/scss/_mixins.scss
@@ -1,0 +1,8 @@
+@mixin placeholder {
+	// sass-lint:disable no-vendor-prefixes
+	&::-webkit-input-placeholder {@content}
+	&:-moz-placeholder {@content}
+	&::-moz-placeholder {@content}
+	&:-ms-input-placeholder {@content}
+	// sass-lint:enable no-vendor-prefixes
+}
