+++ src/scss/email-only-signup.scss
@@ -1,25 +1,3 @@
-@include oColorsSetColor('o-email-only-signup-placeholder-grey', #767676);
-
-@include oColorsSetUseCase(o-email-only-signup-box, border, 'teal-1');
-@include oColorsSetUseCase(o-email-only-signup-input, border, 'teal-1');
-@include oColorsSetUseCase(o-email-only-signup-muted, text, 'grey-tint4');
-@include oColorsSetUseCase(o-email-only-signup-highlight, text, 'teal-1');
-@include oColorsSetUseCase(o-email-only-signup-placeholder, text, 'o-email-only-signup-placeholder-grey');
-
-@mixin placeholder {
-	// sass-lint:disable no-vendor-prefixes
-	&::-webkit-input-placeholder {@content}
-	&:-moz-placeholder {@content}
-	&::-moz-placeholder {@content}
-	&:-ms-input-placeholder {@content}
-	// sass-lint:enable no-vendor-prefixes
-}
-
-.o-email-only-signup__discreet {
-	position: relative;
-	padding-bottom: 15px;
-}
-
 .o-email-only-signup {
 	@include oColorsFor(o-email-only-signup-box, border);
 	position: relative;
@@ -55,6 +33,11 @@
 		padding-right: 15px;
 		border: 1px solid oColorsGetPaletteColor('pink-tint3');
 		background-color: oColorsGetPaletteColor('warm-5');
+	}
+
+	&__discreet {
+		position: relative;
+		padding-bottom: 15px;
 	}
 
 	&__heading {
@@ -118,7 +101,7 @@
 	}
 
 	&__email {
-		@include oFormsCommonField();
+		@include oFormsCommonFieldBase();
 
 		/* custom placeholder color */
 		@include placeholder {
@@ -137,7 +120,7 @@
 	}
 
 	&__select {
-		@include oFormsCommonField();
+		@include oFormsCommonFieldBase();
 		@include oFormsSelect();
 
 		@include placeholder {
